dune-rpc

API

Library dune-rpc

Dune_rpc Implementation of the protocol used by dune rpc. Independent of IO and any specific rpc requests. The protocol described here is stable and is relied on by 3rd party clients.

Library dune-rpc.private

Dune_rpc_private